Evaluate conditions & refactor structure
